+++ source/renderer/app/utils/shelleyLedger.ts
@@ -4,7 +4,13 @@ import {
   AddressType,
   TxAuxiliaryDataType, // CHECK THIS
   StakeCredentialParamsType,
+  CIP36VoteRegistrationFormat,
 } from '@cardano-foundation/ledgerjs-hw-app-cardano';
+import {
+  str_to_path,
+  base58_decode,
+} from '@cardano-foundation/ledgerjs-hw-app-cardano/dist/utils/address';
+import { HexString } from '@cardano-foundation/ledgerjs-hw-app-cardano/dist/types/internal';
 import { encode } from 'borc';
 import blakejs from 'blakejs';
 import _ from 'lodash';
@@ -173,7 +179,7 @@ export function ShelleyTxOutput(
     const addressBuff =
       addressStyle === AddressStyles.ADDRESS_SHELLEY
         ? utils.bech32_decodeAddress(address)
-        : utils.base58_decode(address);
+        : base58_decode(address);
     return encoder.pushAny([addressBuff, coins]);
   }
 
@@ -403,8 +409,7 @@ export const CachedDeriveXpubFactory = (
         lastIndex,
         derivationScheme: derivationScheme.ed25519Mode,
       });
-      // @ts-ignore
-      return utils.hex_to_buf(derivedXpub);
+      return utils.hex_to_buf(derivedXpub as HexString);
     } catch (e) {
       throw e;
     }
@@ -442,7 +447,7 @@ export const prepareLedgerOutput = (
           type: AddressType.BASE_PAYMENT_KEY_STAKE_KEY,
           params: {
             spendingPath: derivationPathToLedgerPath(output.derivationPath),
-            stakingPath: utils.str_to_path("1852'/1815'/0'/2/0"),
+            stakingPath: str_to_path("1852'/1815'/0'/2/0"),
           },
         },
       },
@@ -458,7 +463,7 @@ export const prepareLedgerOutput = (
         addressHex:
           addressStyle === AddressStyles.ADDRESS_SHELLEY
             ? utils.buf_to_hex(utils.bech32_decodeAddress(output.address))
-            : utils.buf_to_hex(utils.base58_decode(output.address)),
+            : utils.buf_to_hex(base58_decode(output.address)),
       },
     },
     amount: output.amount.quantity.toString(),
@@ -469,17 +474,20 @@ export const prepareLedgerAuxiliaryData = (
   txAuxiliaryData: TxAuxiliaryData
 ) => {
   const { votingPubKey, rewardDestinationAddress, type } = txAuxiliaryData;
-
   if (type === CATALYST_VOTING_REGISTRATION_TYPE) {
     return {
-      type: TxAuxiliaryDataType.CATALYST_REGISTRATION,
+      type: TxAuxiliaryDataType.CIP36_REGISTRATION,
       params: {
-        votingPublicKeyHex: votingPubKey,
+        format: CIP36VoteRegistrationFormat.CIP_15,
+        voteKeyHex: votingPubKey,
         stakingPath: rewardDestinationAddress.stakingPath,
-        rewardsDestination: {
-          type: AddressType.REWARD_KEY,
+        paymentDestination: {
+          type: TxOutputDestinationType.DEVICE_OWNED,
           params: {
-            stakingPath: rewardDestinationAddress.stakingPath,
+            type: AddressType.REWARD_KEY,
+            params: {
+              stakingPath: rewardDestinationAddress.stakingPath,
+            },
           },
         },
         nonce: `${txAuxiliaryData.nonce}`,