+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## vNext
 
+### Chores
+
+- Updated `@cardano-foundation/ledgerjs-hw-app-cardano` to version `6.0.0` ([PR 3093](https://github.com/input-output-hk/daedalus/pull/3093))
+
 ## 5.2.0
 
 ### Features